Refactors API structure with controllers and services (#2)
* chore (model): adds product model with validation

Introduces the Product model with attributes for ID, name, and price.

Implements validation to ensure name is present and price is greater than zero.

Includes a method to convert the product to a hash.

* chore (service): adds basic controllers and setup

Adds a base controller with common response methods.
Introduces health check and products controllers.
The product controller contains create, list, and find endpoints.

* chore (mw): moves JSON body parser to middleware

Relocates the JSON body parser to the middleware directory.

This change improves the application's structure by organizing middleware components in a dedicated directory.

* chore (controller): adds base controller and health endpoint

Sets up a base controller with standard response formats.

Introduces a health endpoint to provide basic service status
information.

Creates a product controller and service to handle the creation,
retrieval, and listing of products.

* refactor (config rutes): improves API structure and adds health endpoint

Restructures the API to use controllers for handling requests.

Introduces a health endpoint for monitoring service availability.

Replaces the monolithic product handling logic with a dedicated
`ProductsController`.

Provides a basic landing page with API information and a 404 handler
with suggested endpoints.

Removes the old product handling logic directly in the config.ru file.